cognitive_load_balancer/requirements.txt
pynput==1.7.6
GitPython==3.1.41
google-generativeai==0.3.2
python-dotenv==1.0.1
